[Databases] extract CacheWrapper
