+++ octobot_commons/databases/__init__.py
@@ -23,6 +23,7 @@ from octobot_commons.databases import relational_databases
 
 from octobot_commons.databases import run_databases_identifier
 from octobot_commons.databases import cache_manager
+from octobot_commons.databases import databases_util
 
 from octobot_commons.databases.database_caches import (
     GenericDatabaseCache,
@@ -60,6 +61,10 @@ from octobot_commons.databases.cache_manager import (
     CacheManager,
 )
 
+from octobot_commons.databases.databases_util import (
+    CacheWrapper,
+)
+
 
 __all__ = [
     "GenericDatabaseCache",
@@ -77,4 +82,5 @@ __all__ = [
     "SQLiteDatabase",
     "RunDatabasesIdentifier",
     "CacheManager",
+    "CacheWrapper",
 ]
