+++ octobot_commons/databases/bases/base_database.py
@@ -69,25 +69,30 @@ class BaseDatabase:
         return f"{self.__class__.__name__}, database: {self._database}"
 
     @classmethod
-    @contextlib.asynccontextmanager
-    async def database(cls, *args, with_lock=False, cache_size=None, database_adaptor=None, **kwargs):
-        if with_lock:
+    def _create_database(cls, *args, required_adaptor=False, cache_size=None, database_adaptor=None, **kwargs):
+        if required_adaptor:
             adaptor = kwargs.pop("database_adaptor", adaptors.TinyDBAdaptor)
             if adaptor is None:
                 raise RuntimeError("database_adaptor parameter required")
             adaptor_instance = adaptor(*args, cache_size=cache_size, **kwargs)
-            database = cls(*args, database_adaptor=database_adaptor, cache_size=cache_size, **kwargs)
+            return cls(*args, database_adaptor=database_adaptor, cache_size=cache_size, **kwargs), adaptor_instance
+        return cls(*args, cache_size=cache_size, **kwargs), None
+
+    @classmethod
+    @contextlib.asynccontextmanager
+    async def database(cls, *args, with_lock=False, cache_size=None, database_adaptor=None, **kwargs):
+        database, adaptor_instance = cls._create_database(*args, required_adaptor=with_lock, cache_size=cache_size,
+                                                          database_adaptor=database_adaptor, **kwargs)
+        if with_lock:
             async with document_database.DocumentDatabase.locked_database(adaptor_instance) as db:
                 database._database = db
                 yield database
+                # context manager is taking care of closing the database
             return
-        database = None
         try:
-            database = cls(*args, cache_size=cache_size, **kwargs)
             yield database
         finally:
-            if database is not None:
-                await database.close()
+            await database.close()
 
     @staticmethod
     def get_serializable_value(value):
